Fix failing backups due to no permissions on redis config
As of https://github.com/NixOS/nixpkgs/pull/200319, the redis config is
copied to the redis runtime directory with 0600 permissions. Since
changing these permissions would not only require a significant amount
of work, but the config file does also not contain anything important.
So limiting on the redis dump seems sensible